% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continent_data}
\alias{continent_data}
\title{Continent data}
\format{A data.frame of 3,666 observations
\describe{
  \item{\code{from}}{ edge source}
  \item{\code{to}}{ edge target}
  \item{\code{value}}{ value of edge}
}}
\usage{
continent_data
}
\description{
Random continent data for example and testing purposes.
}
\keyword{datasets}
