% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_data}
\alias{country_data}
\title{Sample data}
\format{A data.frame of 3,666 observations
\describe{
  \item{\code{from}}{ edge source}
  \item{\code{to}}{ edge target}
  \item{\code{value}}{ value of edge}
}}
\source{
\url{https://github.com/syt123450/giojs/blob/master/examples/data/sampleData.json}
}
\usage{
country_data
}
\description{
Random data for example and testing purposes.
}
\keyword{datasets}
