% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{g_clear_p}
\alias{g_clear_p}
\title{Clear}
\usage{
g_clear_p(proxy)
}
\description{
Clear all data
}
\examples{
if(interactive()){

  library(shiny)

  data("country_data")

  ui <- fluidPage(
    actionButton("clear", "clear"),
    giorOutput("gior")
  )

  server <- function(input, output, session){
    output$gior <- renderGior({
      country_data \%>\%
        gior() \%>\%
        g_data(from, to, value)
    })

    observeEvent(input$clear, {
      giorProxy("gior") \%>\%
        g_clear_p()
    })
  }

  shinyApp(ui, server)

}

}
