% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{g_data_p}
\alias{g_data_p}
\title{Add data}
\usage{
g_data_p(proxy, data, from, to, value, ...)
}
\description{
Add data to visualisation.
}
\examples{
if(interactive()){

  library(shiny)

  data("country_data")

  ui <- fluidPage(
    actionButton("add", "add"),
    actionButton("clear", "clear"),
    giorOutput("gior")
  )

  server <- function(input, output, session){
    output$gior <- renderGior({
      data \%>\%
        gior() \%>\%
        g_data(from, to, value)
    })

    observeEvent(input$clear, {
      giorProxy("gior") \%>\%
        g_clear_p()
    })

    observeEvent(input$add, {
      giorProxy("gior") \%>\%
        g_data_p(country_data, from, to, value)
    })
  }

  shinyApp(ui, server)

}

}
