% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{g_init}
\alias{g_init}
\title{Initial Country}
\usage{
g_init(g, country)
}
\arguments{
\item{g}{An object of class \code{gior} as returned by \code{\link{gior}}.}

\item{country}{An ISO code of the country displayed at the center of the screen when the globe has been initialized.}
}
\description{
Set initially selected country
}
\examples{
data("country_data")

country_data \%>\%
  gior() \%>\%
  g_data(from, to, value) \%>\%
  g_init("RU")

}
