% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{g_style}
\alias{g_style}
\title{Style}
\usage{
g_style(g, style)
}
\arguments{
\item{g}{An object of class \code{gior} as returned by \code{\link{gior}}.}

\item{style}{Style, see details.}
}
\description{
Set globe style.
}
\details{
Valid \code{style} include:

\itemize{
  \item{\code{blueInk}}
  \item{\code{earlySpring}}
  \item{\code{frozenBerry}}
  \item{\code{gorgeousDream}}
  \item{\code{juicyCake}}
  \item{\code{lemonGate}}
  \item{\code{magic}}
  \item{\code{mint}}
  \item{\code{nearMoon}}
  \item{\code{octoberParty}}
  \item{\code{redBlue}}
  \item{\code{strawberry}}
  \item{\code{sunset}}
}
}
\examples{
data("country_data")

globe <- country_data \%>\%
  gior(init.country = "CN") \%>\%
  g_data(from, to, value)

g_style(globe, "strawberry")
g_style(globe, "magic")

}
