% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{g_switch_p}
\alias{g_switch_p}
\title{Switch}
\usage{
g_switch_p(proxy, country)
}
\arguments{
\item{proxy}{A proxy as returned by \code{\link{giorProxy}}.}

\item{country}{Country to zoom onto.}
}
\description{
Switch country.
}
