% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gior.R
\name{gior}
\alias{gior}
\title{Initialise}
\usage{
gior(data = NULL, init.country = "US", width = "100\%",
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{init.country}{An ISO code of the country displayed at the center of the screen when the globe has been initialized.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Id of element, auto-generated if \code{NULL}.}
}
\description{
Initialise globe
}
\examples{
data("country_data")

country_data \%>\%
  gior(init.country = "CN") \%>\%
  g_data(from, to, value)

}
