% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{g_init}
\alias{g_init}
\title{Initial Country}
\usage{
g_init(g, country)
}
\arguments{
\item{g}{An object of class \code{gior} as returned by \code{\link{gior}}.}

\item{country}{ISO code 2 of country to have initially selected.}
}
\description{
Set initially selected country
}
