% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gior.R
\name{gior}
\alias{gior}
\title{Initialise}
\usage{
gior(data = NULL, initCountry = "US", width = "100\%",
  height = NULL, elementId = NULL)
}
\description{
Initialise globe
}
\examples{
data("country_data")

country_data \%>\%
  gior() \%>\%
  g_data(from, to, value)

}
